requirements.txt
Django==1.11.15
ortools
beautifulsoup4==4.6.0
requests
Pillow
django-admin-rangefilter
